fixed minor code errors and simplified typos

+++ src/core/pathinfo.c
@@ -15,9 +15,11 @@
 #endif
 
 #ifdef __APPLE__
+
 #include <libgen.h>
 #include <limits.h>
 #include <mach-o/dyld.h>
+
 #endif
 
 #if defined __linux__ || defined __CYGWIN__ || defined __MINGW32_MAJOR_VERSION
@@ -65,16 +67,18 @@ char *get_exe_path() {
 #endif
 
 #ifdef __APPLE__
+
 char *get_exe_path() {
   char raw_path[PATH_MAX];
   char *real_path = malloc(PATH_MAX * sizeof(char));
-  uint32_t raw_size = (uint32_t)sizeof(raw_path);
+  uint32_t raw_size = (uint32_t) sizeof(raw_path);
 
   if (!_NSGetExecutablePath(raw_path, &raw_size)) {
     realpath(raw_path, real_path);
   }
   return real_path;
 }
+
 #endif
 
 #if defined __CYGWIN__ || defined __linux__ || defined __APPLE__
@@ -87,6 +91,7 @@ char *get_exe_dir() {
   free(exe_path_str);
   return exe_dir;
 }
+
 #endif
 
 char *merge_paths(char *a, char *b) {
@@ -112,10 +117,11 @@ char *merge_paths(char *a, char *b) {
 bool file_exists(char *filepath) { return access(filepath, F_OK) == 0; }
 
 #ifndef _WIN32
+
 char *get_calling_dir() { return getenv("PWD"); }
 
 char *get_filename(char *filepath) {
-  int start = 0, length = strlen(filepath);
+  int start = 0, length = (int)strlen(filepath);
   for (int i = 0; i < length; i++) {
     if (filepath[i] == BIRD_PATH_SEPARATOR[0])
       start = i;
@@ -125,6 +131,7 @@ char *get_filename(char *filepath) {
   strncat(string, filepath + start, length);
   return string;
 }
+
 #else
 char *get_calling_dir() { return _fullpath(NULL, "", 0); }
 
@@ -144,7 +151,7 @@ char *resolve_import_path(char *module_name, const char *current_file) {
   char *bird_file_name = get_bird_filename(module_name);
 
   // check relative to the current file...
-  char *file_directory = dirname((char *)current_file);
+  char *file_directory = dirname((char *) current_file);
   char *relative_file = merge_paths(file_directory, bird_file_name);
 
   if (file_exists(relative_file)) {
@@ -152,7 +159,7 @@ char *resolve_import_path(char *module_name, const char *current_file) {
     char *path1 = realpath(relative_file, NULL);
     char *path2 = realpath(current_file, NULL);
 
-    if (path2 == NULL || memcmp(path1, path2, (int)strlen(path2)) != 0)
+    if (path2 == NULL || memcmp(path1, path2, (int) strlen(path2)) != 0)
       return relative_file;
   }
 
@@ -165,7 +172,7 @@ char *resolve_import_path(char *module_name, const char *current_file) {
     char *path1 = realpath(library_file, NULL);
     char *path2 = realpath(current_file, NULL);
 
-    if (path2 == NULL || memcmp(path1, path2, (int)strlen(path2)) != 0)
+    if (path2 == NULL || memcmp(path1, path2, (int) strlen(path2)) != 0)
       return library_file;
   }
 
@@ -173,13 +180,14 @@ char *resolve_import_path(char *module_name, const char *current_file) {
 }
 
 #include <stdio.h>
+
 bool is_core_library_file(char *filepath, char *file_name) {
   char *bird_file_name = get_bird_filename(file_name);
   char *bird_directory = merge_paths(get_exe_dir(), LIBRARY_DIRECTORY);
   char *library_file = merge_paths(bird_directory, bird_file_name);
 
   if (file_exists(library_file)) {
-    return memcmp(library_file, filepath, (int)strlen(filepath)) == 0;
+    return memcmp(library_file, filepath, (int) strlen(filepath)) == 0;
   }
   return false;
 }
